requirements.txt
bcrypt==3.1.4
beautifulsoup4==4.9.1
blinker==1.4
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
click==6.7
elasticsearch==7.7.0
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
html5lib==1.0.1
idna==2.9
itsdangerous==0.24
Jinja2==2.10
lxml==4.5.1
MarkupSafe==1.0
numpy==1.18.4
pandas==1.0.3
pbr==3.1.1
Pillow==5.3.0
pycparser==2.18
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.11.0
soupsieve==2.0.1
SQLAlchemy==1.2.7
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.1
